More optimization: Allocate all LinkNode need for loops in one step, instead of doing one alloc per loop.
This allocate a bit more memory than needed (since we won't use linknodes of 'sharp' loops),
but avoids calling BLI_linklist_prepend_memarena() once per 'smooth' loop.

Note: gave much less benefits than expected, in the end - guess memarena allocation is *really* fast!

+++ source/blender/modifiers/intern/MOD_setsplitnormal.c
@@ -200,7 +200,7 @@ static float vertex_weight(MDeformVert *dvert, const int index, const int defgrp
 
 static void setSplitNormalModifier_do(SetSplitNormalModifierData *smd, Object *ob, DerivedMesh *dm)
 {
-	float (*clnors)[2];
+	short (*clnors)[2];
 	const int num_verts = dm->getNumVerts(dm);
 	const int num_edges = dm->getNumEdges(dm);
 	const int num_loops = dm->getNumLoops(dm);
